Rename classes/tasks in base.
