+++ python/lsst/faro/base/CatalogAnalysisBase.py
