+++ python/lsst/faro/base/CatalogSummaryBase.py
@@ -5,11 +5,11 @@ import lsst.pex.config as pexConfig
 from .BaseSubTasks import NumpyAggTask
 
 
-# Dimentions of the Connections class define the iterations of runQuantum
-class CatalogsAggregationBaseTaskConnections(MetricConnections,
-                                             defaultTemplates={'agg_name': None},
-                                             dimensions=("band", "tract",
-                                                         "instrument", "skymap")):
+# Dimensions of the Connections class define the iterations of runQuantum
+class CatalogSummaryBaseTaskConnections(MetricConnections,
+                                        defaultTemplates={'agg_name': None},
+                                        dimensions=("band", "tract",
+                                                    "instrument", "skymap")):
     # Make this an LSST verify Measurement
     measurement = pipeBase.connectionTypes.Output(doc="{agg_name} {package}_{metric}.",
                                                   dimensions=("instrument", "tract", "band"),
@@ -17,8 +17,8 @@ class CatalogsAggregationBaseTaskConnections(MetricConnections,
                                                   name="metricvalue_{agg_name}_{package}_{metric}")
 
 
-class CatalogAggregationBaseTaskConfig(MetricConfig,
-                                       pipelineConnections=CatalogsAggregationBaseTaskConnections):
+class CatalogSummaryBaseTaskConfig(MetricConfig,
+                                   pipelineConnections=CatalogsSummaryBaseTaskConnections):
     agg = pexConfig.ConfigurableField(
         # This task is meant to make measurements of various types.
         # The default task is, therefore, a bit of a place holder.
@@ -28,10 +28,10 @@ class CatalogAggregationBaseTaskConfig(MetricConfig,
         doc="Numpy aggregation task")
 
 
-class CatalogsAggregationBaseTask(MetricTask):
+class CatalogsSummaryBaseTask(MetricTask):
 
-    ConfigClass = CatalogAggregationBaseTaskConfig
-    _DefaultName = "catalogsAggregationBaseTask"
+    ConfigClass = CatalogSummaryBaseTaskConfig
+    _DefaultName = "catalogSummaryBaseTask"
 
     def __init__(self, config, *args, **kwargs):
         super().__init__(*args, config=config, **kwargs)
